Added routing for track button
Used asp-controller
