+++ Basecode.WebApp/Views/Home/Index.cshtml
@@ -7,7 +7,7 @@
     <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
 
     <p>Already have a job application in progress?</p>
-    <form method="get" asp-page="SearchResults">
+    <form method="get" asp-controller="Home" asp-action="JobsList">
         <input type="text" name="searchQuery" placeholder="Enter applicant ID here" />
         <button type="submit">Track</button>
     </form>
