+++ Basecode.WebApp/Views/Home/SearchResults.cshtml
@@ -1,4 +1,4 @@
-@page
+@* @page
 @model Basecode.Main.Models.SearchResultsModel
 
 <h1>Search Results</h1>
@@ -15,4 +15,4 @@
 else
 {
     <p>No results found.</p>
-}
+} *@
